+++ tests/run.clj
@@ -72,6 +72,7 @@
 (require "tests/spawn_stress_regression")
 (require "tests/tasks_test")
 (require "tests/records_test")
+(require "tests/instant_template_test")
 (require "tests/async_buffer_test")
 (require "tests/async_api_test")
 (require "tests/async_alts_test")